[SWDEV-182]auto-matching-datacenter

+++ dcmgr/scheduler/datacenter.go
@@ -3,8 +3,8 @@ package scheduler
 type DataCenterRecord struct {
 	ID           string
 	Name         string
-	CPU          int
-	Memory       int
-	Disk         int
+	CPU          int64
+	Memory       int64
+	Disk         int64
     Price        float64
 }
